Add 2 new parameters: 'ils-i' and 'min-k'
With the 'ils-i' parameter, user can assign a value for the number of iterations of internal ILS loop.
Actually, we will not use this parameter now, but we may use it later.

With the 'min-k' paramter, ILS chooses the solution with minimal cluster number
when the best ils solution and the current ils solution have the same imbalance value during ILS iterations.

One last thing about 'init-partition-file' and 'i' parameters ('i' means the number of outer ils iterations):
(we did not do any change about this, we write this just for the purpose of remembering)
The option 'i' will be updated with the 'init-partition-file' parameter. The idea is as follows:
If user provides an initial partition, it is more logical and fair that ILS runs only once (1 iteration).
Because, ILS has <i> iterations to find the best solution. In each iteration, the initial partition is generated.
Consequently, the initial partition of the iterations are different.
On the other hand, when user provides ILS with initial partition, the initial partition will be the same for all ILS iterations.